Fix some minor UI issues, Satisfy Linter, Improve a few colors
