+++ app/src/main/java/sensors_in_paradise/sonar/page2/Page2Handler.kt
@@ -22,7 +22,6 @@ class Page2Handler(
     private lateinit var timer: Chronometer
 
     private lateinit var recyclerViewRecordings: RecyclerView
-    private lateinit var activityCountTextView: TextView
     private lateinit var viewSwitcher: ViewSwitcher
     private lateinit var recordingsAdapter: RecordingsAdapter
     private lateinit var tabLayout: TabLayout
@@ -37,8 +36,6 @@ class Page2Handler(
         this.context = activity
         this.activity = activity
         timer = activity.findViewById(R.id.timer)
-
-        activityCountTextView = activity.findViewById(R.id.tv_activity_counts)
 
         recyclerViewRecordings = activity.findViewById(R.id.recyclerView_recordings_captureFragment)
         val linearLayoutManager = LinearLayoutManager(context)
@@ -73,22 +70,11 @@ class Page2Handler(
         }
 
         tabLayout.addOnTabSelectedListener(this)
-        updateActivityCounts()
     }
 
     private fun addRecordingToUI(recording: Recording) {
         recordingsManager.recordingsList.add(0, recording)
         recordingsAdapter.notifyItemInserted(0)
-        updateActivityCounts()
-    }
-
-    private fun updateActivityCounts() {
-        val numberOfRecodings = recordingsManager.getNumberOfRecordingsPerActivity()
-        var text = " "
-        for ((activity, number) in numberOfRecodings) {
-            text += "$activity: $number | "
-        }
-        activityCountTextView.text = text.trimEnd('|', ' ')
     }
 
     override fun activityResumed() {
