+++ app/src/main/java/sensors_in_paradise/sonar/page2/RecordingDataManager.kt
@@ -3,10 +3,8 @@ package sensors_in_paradise.sonar.page2
 import sensors_in_paradise.sonar.GlobalValues
 import java.io.File
 
-class RecordingDataManager(private val recordingsDir: File) {
-    val recordingsList = ArrayList<Recording>()
-
-    init {
+class RecordingDataManager(private val recordingsDir: File, val recordingsList: ArrayList<Recording>) {
+    constructor(recordingsDir: File) : this(recordingsDir, ArrayList<Recording>()) {
         loadRecordingsFromStorage()
     }
