+++ app/src/main/java/sensors_in_paradise/sonar/page3/PredictionsAdapter.kt
@@ -4,11 +4,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import androidx.cardview.widget.CardView
 import androidx.recyclerview.widget.RecyclerView
 import sensors_in_paradise.sonar.R
 
 class PredictionsAdapter(
     private val predictions: ArrayList<Prediction>,
+    private val highlightColor: Int
 ) :
     RecyclerView.Adapter<PredictionsAdapter.ViewHolder>() {
 
@@ -19,6 +21,7 @@ class PredictionsAdapter(
     class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         val titleTextView: TextView = view.findViewById(R.id.tv_title_prediction)
         val detailsTextView: TextView = view.findViewById(R.id.tv_content_prediction)
+        val cardView: CardView = view.findViewById(R.id.card_parent_prediction)
     }
 
     // Create new views (invoked by the layout manager)
@@ -32,12 +35,12 @@ class PredictionsAdapter(
 
     // Replace the contents of a view (invoked by the layout manager)
     override fun onBindViewHolder(viewHolder: ViewHolder, position: Int) {
-
-        // Get element from your dataset at this position and replace the
-        // contents of the view with that element
        val prediction = predictions[position]
         viewHolder.titleTextView.text = prediction.title
         viewHolder.detailsTextView.text = prediction.percentageAsString()
+        if (position == 0) {
+            viewHolder.cardView.setCardBackgroundColor(highlightColor)
+        }
     }
 
     // Return the size of your dataset (invoked by the layout manager)
