+++ app/src/main/res/color/primary_selector.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:color="@color/colorPrimary" android:state_selected="true"/>
-    <item android:color="@color/grey"/>
+    <item android:color="?android:textColorSecondary"/>
 </selector>
