+++ app/src/main/res/drawable-night/tablayout_background.xml
@@ -15,7 +15,7 @@
         <shape android:shape="rectangle">
             <stroke
                 android:width="1dp"
-                android:color="@color/colorPrimary" />
+                android:color="@color/slightBackgroundContrast" />
             <solid android:color="#00000000" />
         </shape>
     </item>
