+++ app/src/main/res/layout/capture_fragment.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:paddingHorizontal="16dp"
-    android:paddingVertical="16dp"
+    android:layout_marginTop="16dp"
     tools:context=".MainActivity"
    >
 
@@ -132,30 +132,21 @@
             android:layout_height="match_parent"
             android:orientation="vertical">
 
-            <TextView
-                android:id="@+id/tv_activity_counts"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="TextView" />
-
             <androidx.recyclerview.widget.RecyclerView
                 android:id="@+id/recyclerView_recordings_captureFragment"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
+                android:clipToPadding="false"
                 app:layoutManager="LinearLayoutManager"/>
         </LinearLayout>
 
-        <LinearLayout
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/recyclerView_activities_captureFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:orientation="vertical">
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/recyclerView_activities_captureFragment"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                app:layoutManager="LinearLayoutManager"/>
-        </LinearLayout>
+            android:clipToPadding="false"
+            android:paddingVertical="10dp"
+            app:layoutManager="LinearLayoutManager"/>
 
     </ViewSwitcher>
 </LinearLayout>
