+++ app/src/main/res/layout/upload_dialog.xml
@@ -13,7 +13,15 @@
         android:text="Sensor data files"
         android:textSize="20sp"
         android:textStyle="bold" />
-
+    <TextView
+        android:id="@+id/tv_activity_counts"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginHorizontal="10dp"
+        android:layout_marginBottom="10dp"
+        android:text="Activity counts"
+        android:visibility="gone"
+         />
     <ViewSwitcher
         android:id="@+id/switcher_connectionStatus_uploadDialog"
         android:layout_width="match_parent"
